# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hamrofitness Api Documentation",
        "description": "Hamrofitness Api Documentation",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/merchant/home": {
            "get": {
                "tags": [
                    "Frontend"
                ],
                "summary": "Get home details",
                "operationId": "ba11752a369581c49380f68fd75f0cad",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/merchant-info/{mobileApp}": {
            "get": {
                "tags": [
                    "Frontend"
                ],
                "summary": "Get merchant information",
                "operationId": "49113ba6822619994d64845dfc15933b",
                "parameters": [
                    {
                        "name": "mobileApp",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/tutorials": {
            "get": {
                "tags": [
                    "Frontend"
                ],
                "summary": "Get tutorial information. Add type: image,text,image,audio,video",
                "operationId": "20c21802dfc51d6168ce6d066dd657f5",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "The type of tutorial (optional)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/payment-sources": {
            "get": {
                "tags": [
                    "Frontend"
                ],
                "summary": "Get Payment Sources information",
                "operationId": "2495b65671b1613b7b6714da4e67eb2c",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport apiKey security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}